linux: moved NewConError to linux/mod.rs and implemented Con::new() and Con::try_default()
